backend/requirements.txt
Django==4.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-filter==22.1